feat: initial dashboard operations

+++ finefin.backend/src/finefin.api/Http/Responses/SummaryResponse.cs
@@ -2,10 +2,10 @@
 {
     public class SummaryResponse
     {
-        public double TotalBalance { get; set; }
-        public double CurrentMonthIncomes { get; set; }
-        public double CurrentMonthExpenses { get; set; }
-        public double CurrentMonthBalance { get; set; }
+        public decimal TotalBalance { get; set; }
+        public decimal CurrentMonthIncomes { get; set; }
+        public decimal CurrentMonthExpenses { get; set; }
+        public decimal CurrentMonthBalance { get; set; }
 
     }
 }
